buildtype debug일때만 로그 찍기 콜백함수 위치 변경
